Allow fractional quantities in the panier and commandes
LigneCommande.quantite was an int, so the quantity stepper (in the
panier and the admin's commande-edit sheet) only accepted whole
numbers. Some articles are sold by non-integer quantity (e.g. 0.4),
so quantite is now a double throughout: the model, Firestore
serialization, the stepper's text field (decimal keyboard, accepts
",", " and "."), and every screen displaying a quantity (facture PDF,
facture preview, commande recap, commande tiles).

Added formatQuantite() to show whole quantities without a trailing
decimal ("10") and fractional ones trimmed to their needed decimals
("0,4"), French-style. The cart badge on the catalogue toolbar now
counts distinct line items instead of summing quantities, since that
sum can no longer be a clean integer.

+++ lib/widgets/commande_edit_sheet.dart
@@ -148,11 +148,11 @@ class _CommandeEditSheetState extends State<_CommandeEditSheet> {
     ),
   );
 
-  void _modifierQuantite(int index, int delta) {
+  void _modifierQuantite(int index, double delta) {
     _definirQuantite(index, _lignes[index].quantite + delta);
   }
 
-  void _definirQuantite(int index, int quantite) {
+  void _definirQuantite(int index, double quantite) {
     setState(() {
       if (quantite <= 0) {
         _lignes.removeAt(index);
